Refactor stats components
